Queue notification (WIP)
Use a queue to better control the flow and to be able to retry if
the APN service is down or similar. Also decouples the APN call
from the API call.

This won't work at the moment since the API caller needs to know
the result from APN. This could be solved with a webhook.