Add authentication to the API
Add authentication to the API to prevent abuse and spam and only
allow push notifications from Pine's server to begin with. The
beta will only allow Pine addresses anyhow and when it opens up
for people to host their own Pine servers, they will have to
register their server and get an API key in order to send push
notifications to the Pine app. If they abuse their API key,
the API key would be revoked.